+++ client/src/app/views/edit-add/edit-add.directives/upload-image/upload-image.directive.js
@@ -60,10 +60,10 @@
                 case 'camera':
                     imgUrl = 'assets/images/camera_45x34.svg';
                     break;
-                case 'misc':
-                    imgUrl = 'assets/images/circle-thin-checkit-green.svg';
+                case 'software':
+                    imgUrl = 'assets/images/code.svg';
                     break;
-                case 'watch':
+                case 'misc':
                     imgUrl = 'assets/images/circle-thin-checkit-green.svg';
                     break;
                 default: